Extract common controller for dissolving titles
