+++ Eurofurence/Application/Components/Dealer Detail/View/UIKit/DealerDetailViewController.swift
@@ -4,19 +4,22 @@ class DealerDetailViewController: UIViewController, DealerDetailScene {
 
     // MARK: Properties
     
-    private lazy var titleView = DissolvingTitleLabel(frame: .zero)
-    private unowned var titleLabelForScrollingTitleUpdates: UILabel?
+    private var titleController: DissolvingTitleController?
 
     @IBOutlet private weak var tableView: UITableView!
-    
-    private var contentOffsetObservation: NSKeyValueObservation?
     
     private var tableController: TableController? {
         didSet {
             tableView.dataSource = tableController
             tableController?.titleAvailable = { [weak self] (titleLabel) in
-                self?.titleLabelForScrollingTitleUpdates = titleLabel
-                self?.titleView.text = titleLabel.text
+                guard let self = self else { return }
+                
+                self.titleController = DissolvingTitleController(
+                    scrollView: self.tableView,
+                    navigationItem: self.navigationItem,
+                    titleLabel: titleLabel,
+                    accessibilityIdentifier: "org.eurofurence.dealer.navigationTitle"
+                )
             }
         }
     }
@@ -25,15 +28,6 @@ class DealerDetailViewController: UIViewController, DealerDetailScene {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        titleView.accessibilityIdentifier = "org.eurofurence.dealer.navigationTitle"
-        navigationItem.titleView = titleView
-        
-        contentOffsetObservation = tableView.observe(\.contentOffset, changeHandler: { [weak self] (_, _) in
-            guard let newValue = self?.tableView.contentOffset else { return }
-            self?.updateNavigationTitle(contentOffset: newValue)
-        })
-        
         delegate?.dealerDetailSceneDidLoad()
     }
     
@@ -45,48 +39,6 @@ class DealerDetailViewController: UIViewController, DealerDetailScene {
     override func viewDidDisappear(_ animated: Bool) {
         super.viewDidDisappear(animated)
         delegate?.dealerDetailSceneDidDisappear()
-    }
-    
-    // MARK: Title management
-    
-    private func updateNavigationTitle(contentOffset: CGPoint) {
-        if case .visible(let opacity) = shouldShowNavigationTitle(contentOffset: contentOffset) {
-            showTitle(opacity: opacity)
-        } else {
-            hideTitle()
-        }
-    }
-    
-    private enum TitleState {
-        case visible(opacity: CGFloat)
-        case hidden
-    }
-    
-    private func shouldShowNavigationTitle(contentOffset: CGPoint) -> TitleState {
-        guard let titleLabelForScrollingTitleUpdates = titleLabelForScrollingTitleUpdates else { return .hidden }
-        
-        let titleLabelFrame = titleLabelForScrollingTitleUpdates.frame
-        let titleLabelTop = titleLabelFrame.origin.y - contentOffset.y
-        let navigationTitleFrame = titleView.frame
-        let additionalPaddingForAnimation: CGFloat = 20
-        let navigationTitleBottom = navigationTitleFrame.origin.y + navigationTitleFrame.size.height + additionalPaddingForAnimation
-        
-        if titleLabelTop < navigationTitleBottom {
-            let opacity = max(0, 1 - (titleLabelTop / navigationTitleBottom))
-            return .visible(opacity: opacity)
-        } else {
-            return .hidden
-        }
-    }
-    
-    private func hideTitle() {
-        titleView.isHidden = true
-        titleView.alpha = 0
-    }
-    
-    private func showTitle(opacity: CGFloat) {
-        titleView.isHidden = false
-        titleView.alpha = opacity
     }
     
     // MARK: Actions
